refactor indexer and merge indexer struct define
