+++ crates/rooch-indexer/src/actor/indexer.rs
@@ -4,17 +4,18 @@
 use crate::actor::messages::{
     IndexerEventsMessage, IndexerStatesMessage, IndexerTransactionMessage,
 };
-use crate::actor::{new_field_state, new_object_state_from_raw_object};
 use crate::store::traits::IndexerStoreTrait;
-use crate::types::{IndexedEvent, IndexedFieldState, IndexedObjectState, IndexedTransaction};
 use crate::IndexerStore;
 use anyhow::Result;
 use async_trait::async_trait;
 use coerce::actor::{context::ActorContext, message::Handler, Actor};
-use move_core_types::effects::Op;
 use moveos_store::MoveOSStore;
-use moveos_types::moveos_std::object::{ObjectID, RootObjectEntity};
-use moveos_types::state::{FieldChange, ObjectChange};
+use moveos_types::moveos_std::object::RootObjectEntity;
+use rooch_types::indexer::event::IndexerEvent;
+use rooch_types::indexer::state::{
+    handle_object_change, IndexerFieldStateChanges, IndexerObjectStateChanges,
+};
+use rooch_types::indexer::transaction::IndexerTransaction;
 
 pub struct IndexerActor {
     root: RootObjectEntity,
@@ -34,95 +35,6 @@ impl IndexerActor {
             _moveos_store: moveos_store,
         })
     }
-
-    //TODO wrap the arguments to a struct
-    #[allow(clippy::too_many_arguments)]
-    fn handle_object_change(
-        mut state_index_generator: u64,
-        tx_order: u64,
-        new_object_states: &mut Vec<IndexedObjectState>,
-        update_object_states: &mut Vec<IndexedObjectState>,
-        remove_object_states: &mut Vec<String>,
-        remove_field_states_by_object_id: &mut Vec<String>,
-        new_field_states: &mut Vec<IndexedFieldState>,
-        update_field_states: &mut Vec<IndexedFieldState>,
-        remove_field_states: &mut Vec<(String, String)>,
-        object_id: ObjectID,
-        object_change: ObjectChange,
-    ) -> Result<u64> {
-        let ObjectChange { op, fields } = object_change;
-
-        if let Some(op) = op {
-            match op {
-                Op::Modify(value) => {
-                    debug_assert!(value.is_object());
-                    let state =
-                        new_object_state_from_raw_object(value, tx_order, state_index_generator)?;
-                    update_object_states.push(state);
-                }
-                Op::Delete => {
-                    remove_object_states.push(object_id.to_string());
-                    remove_field_states_by_object_id.push(object_id.to_string());
-                }
-                Op::New(value) => {
-                    debug_assert!(value.is_object());
-                    let state =
-                        new_object_state_from_raw_object(value, tx_order, state_index_generator)?;
-                    new_object_states.push(state);
-                }
-            }
-        }
-
-        state_index_generator += 1;
-        for (key, change) in fields {
-            match change {
-                FieldChange::Normal(normal_change) => {
-                    match normal_change.op {
-                        Op::Modify(value) => {
-                            let state = new_field_state(
-                                key,
-                                value,
-                                object_id.clone(),
-                                tx_order,
-                                state_index_generator,
-                            )?;
-                            update_field_states.push(state);
-                        }
-                        Op::Delete => {
-                            remove_field_states.push((object_id.to_string(), key.to_string()));
-                        }
-                        Op::New(value) => {
-                            let state = new_field_state(
-                                key,
-                                value,
-                                object_id.clone(),
-                                tx_order,
-                                state_index_generator,
-                            )?;
-                            new_field_states.push(state);
-                        }
-                    }
-                    state_index_generator += 1;
-                }
-                FieldChange::Object(object_change) => {
-                    state_index_generator = Self::handle_object_change(
-                        state_index_generator,
-                        tx_order,
-                        new_object_states,
-                        update_object_states,
-                        remove_object_states,
-                        remove_field_states_by_object_id,
-                        new_field_states,
-                        update_field_states,
-                        remove_field_states,
-                        key.as_object_id()?,
-                        object_change,
-                    )?;
-                }
-            }
-        }
-        Ok(state_index_generator)
-    }
 }
 
 impl Actor for IndexerActor {}
@@ -140,49 +52,38 @@ impl Handler<IndexerStatesMessage> for IndexerActor {
 
         // indexer state index generator
         let mut state_index_generator = 0u64;
-        let mut new_object_states = vec![];
-        let mut update_object_states = vec![];
-        let mut remove_object_states = vec![];
-
-        let mut new_field_states = vec![];
-        let mut update_field_states = vec![];
-        let mut remove_field_states = vec![];
-
-        // When remove table handle, first delete table handle from global states,
-        // then delete all states which belongs to the object_id from table states
-        let mut remove_field_states_by_object_id = vec![];
+        let mut indexer_object_state_changes = IndexerObjectStateChanges::default();
+        let mut indexer_field_state_changes = IndexerFieldStateChanges::default();
 
         for (object_id, object_change) in state_change_set.changes {
-            state_index_generator = Self::handle_object_change(
+            state_index_generator = handle_object_change(
                 state_index_generator,
                 tx_order,
-                &mut new_object_states,
-                &mut update_object_states,
-                &mut remove_object_states,
-                &mut remove_field_states_by_object_id,
-                &mut new_field_states,
-                &mut update_field_states,
-                &mut remove_field_states,
+                &mut indexer_object_state_changes,
+                &mut indexer_field_state_changes,
                 object_id,
                 object_change,
             )?;
         }
 
         //Merge new object states and update object states
-        new_object_states.append(&mut update_object_states);
+        let mut object_states_new_and_update = indexer_object_state_changes.new_object_states;
+        object_states_new_and_update.append(&mut indexer_object_state_changes.update_object_states);
         self.indexer_store
-            .persist_or_update_object_states(new_object_states)?;
+            .persist_or_update_object_states(object_states_new_and_update)?;
         self.indexer_store
-            .delete_object_states(remove_object_states)?;
+            .delete_object_states(indexer_object_state_changes.remove_object_states)?;
 
         //Merge new field states and update field states
-        new_field_states.append(&mut update_field_states);
+        let mut fiels_states_new_and_update = indexer_field_state_changes.new_field_states;
+        fiels_states_new_and_update.append(&mut indexer_field_state_changes.update_field_states);
         self.indexer_store
-            .persist_or_update_field_states(new_field_states)?;
+            .persist_or_update_field_states(fiels_states_new_and_update)?;
         self.indexer_store
-            .delete_field_states(remove_field_states)?;
-        self.indexer_store
-            .delete_field_states_by_object_id(remove_field_states_by_object_id)?;
+            .delete_field_states(indexer_field_state_changes.remove_field_states)?;
+        self.indexer_store.delete_field_states_by_object_id(
+            indexer_field_state_changes.remove_field_states_by_object_id,
+        )?;
 
         Ok(())
     }
@@ -201,8 +102,8 @@ impl Handler<IndexerTransactionMessage> for IndexerActor {
             moveos_tx,
         } = msg;
 
-        let indexed_transaction = IndexedTransaction::new(transaction, execution_info, moveos_tx)?;
-        let transactions = vec![indexed_transaction];
+        let indexer_transaction = IndexerTransaction::new(transaction, execution_info, moveos_tx)?;
+        let transactions = vec![indexer_transaction];
 
         // just for bitcoin block data import, don't write transaction indexer
         self.indexer_store.persist_transactions(transactions)?;
@@ -221,7 +122,7 @@ impl Handler<IndexerEventsMessage> for IndexerActor {
 
         let events: Vec<_> = events
             .into_iter()
-            .map(|event| IndexedEvent::new(event, transaction.clone(), moveos_tx.clone()))
+            .map(|event| IndexerEvent::new(event, transaction.clone(), moveos_tx.clone()))
             .collect();
         self.indexer_store.persist_events(events)?;
         Ok(())
