+++ crates/rooch-indexer/src/actor/messages.rs
@@ -7,11 +7,11 @@ use moveos_types::moveos_std::event::Event;
 use moveos_types::moveos_std::object::RootObjectEntity;
 use moveos_types::state::StateChangeSet;
 use moveos_types::transaction::{TransactionExecutionInfo, VerifiedMoveOSTransaction};
-use rooch_types::indexer::event_filter::{EventFilter, IndexerEvent, IndexerEventID};
+use rooch_types::indexer::event::{EventFilter, IndexerEvent, IndexerEventID};
 use rooch_types::indexer::state::{
     FieldStateFilter, IndexerFieldState, IndexerObjectState, IndexerStateID, ObjectStateFilter,
 };
-use rooch_types::indexer::transaction_filter::TransactionFilter;
+use rooch_types::indexer::transaction::TransactionFilter;
 use rooch_types::transaction::LedgerTransaction;
 use rooch_types::transaction::TransactionWithInfo;
 use serde::{Deserialize, Serialize};
