+++ crates/rooch-indexer/src/models/events.rs
@@ -1,7 +1,6 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::types::IndexedEvent;
 use crate::{schema::events, utils};
 use diesel::prelude::*;
 use move_core_types::account_address::AccountAddress;
@@ -9,7 +8,7 @@ use move_core_types::language_storage::StructTag;
 use moveos_types::h256::H256;
 use moveos_types::moveos_std::event::EventID;
 use moveos_types::moveos_std::object::ObjectID;
-use rooch_types::indexer::event_filter::{IndexerEvent, IndexerEventID};
+use rooch_types::indexer::event::{IndexerEvent, IndexerEventID};
 use std::str::FromStr;
 
 #[derive(Queryable, QueryableByName, Insertable, Debug, Clone)]
@@ -45,17 +44,17 @@ pub struct StoredEvent {
     pub created_at: i64,
 }
 
-impl From<IndexedEvent> for StoredEvent {
-    fn from(event: IndexedEvent) -> Self {
+impl From<IndexerEvent> for StoredEvent {
+    fn from(event: IndexerEvent) -> Self {
         Self {
-            event_handle_id: event.event_handle_id.to_string(),
-            event_seq: event.event_seq as i64,
+            event_handle_id: event.event_id.event_handle_id.to_string(),
+            event_seq: event.event_id.event_seq as i64,
             event_type: utils::format_struct_tag(&event.event_type),
             event_data: event.event_data,
-            event_index: event.event_index as i64,
+            event_index: event.indexer_event_id.event_index as i64,
 
             tx_hash: format!("{:?}", event.tx_hash),
-            tx_order: event.tx_order as i64,
+            tx_order: event.indexer_event_id.tx_order as i64,
             sender: event.sender.to_hex_literal(),
             created_at: event.created_at as i64,
         }
