+++ crates/rooch-rpc-api/src/jsonrpc_types/rooch_types.rs
@@ -12,7 +12,7 @@ use crate::jsonrpc_types::{
 };
 use move_core_types::u256::U256;
 use rooch_types::framework::coin::CoinInfo;
-use rooch_types::indexer::event_filter::IndexerEventID;
+use rooch_types::indexer::event::IndexerEventID;
 use rooch_types::indexer::state::IndexerStateID;
 use rooch_types::transaction::rooch::RoochTransaction;
 use schemars::JsonSchema;
