+++ crates/rooch-types/src/indexer/mod.rs
@@ -1,8 +1,9 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
-pub mod event_filter;
+
+pub mod event;
 pub mod state;
-pub mod transaction_filter;
+pub mod transaction;
 
 pub trait Filter<T> {
     fn matches(&self, item: &T) -> bool;
