--- crates/rooch-types/src/indexer/transaction_filter.rs
@@ -1,31 +1,0 @@
-// Copyright (c) RoochNetwork
-// SPDX-License-Identifier: Apache-2.0
-
-use move_core_types::account_address::AccountAddress;
-use moveos_types::h256::H256;
-use serde::{Deserialize, Serialize};
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub enum TransactionFilter {
-    /// Query by sender address.
-    Sender(AccountAddress),
-    /// Query by multi chain original address.
-    OriginalAddress(String),
-    /// Query by the transaction hash list.
-    TxHashes(Vec<H256>),
-    /// Return transactions in [start_time, end_time) interval
-    TimeRange {
-        /// left endpoint of time interval, milliseconds since epoch, inclusive
-        start_time: u64,
-        /// right endpoint of time interval, milliseconds since epoch, exclusive
-        end_time: u64,
-    },
-    /// Return transactions in [from_order, to_order) interval
-    TxOrderRange {
-        /// left endpoint of transaction order, inclusive
-        from_order: u64,
-        /// right endpoint of transaction order, exclusive
-        to_order: u64,
-    },
-}
