+++ docs/website/pages/blog/rooch-btc.en-US.mdx
@@ -76,8 +76,8 @@ At Rooch, we would like to propose a new type of solution that leverages an on-c
 
 - Rooch has the flexibility to be customized into any infra architectures, and in this specific case, we will employ Ethereum and an external DA to guarantee security.
 - On top of Rooch, a Bitcoin light client is implemented using smart contracts, which has the ability to verify Bitcoin block heads.
-- All Bitcoin block heads and UTXO states are synced to Rooch as objects in the MoveVM, where the UTXOs are indexed and cross verified by the light client and all the block heads.
-- The indexed Bitcoin UTXO information can then be used in the smart contracts of the fully on-chain application built on Rooch.
+- All Bitcoin block heads and UTXO states are synced to Rooch as objects in the MoveVM, where the UTXOs are indexer and cross verified by the light client and all the block heads.
+- The indexer Bitcoin UTXO information can then be used in the smart contracts of the fully on-chain application built on Rooch.
 
 We will also provide interpreter contracts on Rooch for applications to access different inscription information easily, such as Ordinals, BRC20, ARC20, and more.
 
@@ -136,7 +136,7 @@ Rooch is a modular blockchain platform that aims to support fully on-chain appli
 
 Although we are yet ready to launch the full version of mainnet, in the sight of the recent market demand, we have decided to soon launch a pre-mainnet, the **RoochBTC,** in the hope of showing the community what we’ve been working on and get you guys some hands-on experience.
 
-The RoochBTC network will have the full Bitcoin mainnet state indexed on the chain, as well as everything you need for developing a fully on-chain application with Bitcoin assets. This includes SDK, wallet API, documents, and examples.
+The RoochBTC network will have the full Bitcoin mainnet state indexer on the chain, as well as everything you need for developing a fully on-chain application with Bitcoin assets. This includes SDK, wallet API, documents, and examples.
 
 We will also release demo applications for Bitcoin asset holders to play with and get a taste of the future of Bitcoin fully on-chain applications.
