+++ docs/website/pages/learn/portal.en-US.mdx
@@ -104,7 +104,7 @@ NFT displays our common non-fungible assets.
 
 #### Indexed Assets
 
-Indexed Assets displays some indexed assets, such as UTXO.
+Indexed Assets displays some indexer assets, such as UTXO.
 
 > Coming Soon!
