+++ sdk/circomlib/circuits/sha256partial.circom
@@ -89,7 +89,7 @@ template Sha256Partial(maxBitsPadded) {
         for (j=0; j<maxBlocks; j++) {
             arraySelectors[k].in[j] <== sha256compression[j].out[k];
         }
-        arraySelectors[k].index <== inBlockIndex - 1; // The index is 0 indexed and the block numbers are 1 indexed.
+        arraySelectors[k].index <== inBlockIndex - 1; // The index is 0 indexer and the block numbers are 1 indexer.
         out[k] <== arraySelectors[k].out;
     }
